updating core to no emission based protocol

+++ contracts/helpers/BitShift.sol
@@ -7,7 +7,12 @@ library BitShift {
     function bitShift(
         uint256[] memory tickets, 
         uint256[] memory amountPerTicket
-    ) internal pure returns(uint256 comTickets, uint256 comAmounts, uint256 largestTicket, uint256 base) {
+    ) internal pure returns(
+        uint256 comTickets, 
+        uint256 comAmounts, 
+        uint256 largestTicket, 
+        uint128 base
+    ) {
         uint256 length = tickets.length;
         for(uint256 i = 0; i < length; i++) {
             if(tickets[i] > largestTicket) largestTicket = tickets[i];
@@ -15,7 +20,7 @@ library BitShift {
             comAmounts <<= 25;
             comTickets |= tickets[i];
             comAmounts |= amountPerTicket[i] * 100;
-            base += amountPerTicket[i] * 0.001 ether;
+            base += uint128(amountPerTicket[i] * 0.001 ether);
         }
     }
 }
